backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
pytest==8.3.5
httpx==0.27.2
alembic==1.14.1